refactor: rename database files
database_helper.py -> database.py
database.py -> repository.py
